Validating and Fixing Bugs

+++ admin-panel/jobs-admins/delete-jobs.php
@@ -1,6 +1,9 @@
 <?php require "../layouts/header.php"; ?>    
 <?php require "../../config/config.php"; ?>
 <?php
+if(!isset($_SESSION['email'])){
+    header("location:".ADMINURL."/admins/login-admins.php");
+   }
     if(isset($_GET['id'])){
         $id = $_GET['id'];
         $delete = $conn->prepare("DELETE FROM `jobs` WHERE id='$id'");
